cmd/bosun: move ping out of sched
fixes #1880
