+++ cmd/bosun/main.go
@@ -20,6 +20,7 @@ import (
 
 	"bosun.org/cmd/bosun/conf"
 	"bosun.org/cmd/bosun/conf/rule"
+	"bosun.org/cmd/bosun/ping"
 	"bosun.org/cmd/bosun/sched"
 	"bosun.org/cmd/bosun/web"
 	"bosun.org/collect"
@@ -146,6 +147,9 @@ func main() {
 			sysProvider.SetTSDBHost(tsdbHost.Host)
 		}
 	}
+	if systemConf.GetPing() {
+		go ping.PingHosts(sched.DefaultSched.Search, systemConf.GetPingDuration())
+	}
 	if sysProvider.GetInternetProxy() != "" {
 		web.InternetProxy, err = url.Parse(sysProvider.GetInternetProxy())
 		if err != nil {
@@ -213,6 +217,7 @@ func main() {
 			sched.Run()
 		}
 	}()
+
 	go func() {
 		sc := make(chan os.Signal, 1)
 		signal.Notify(sc, os.Interrupt, syscall.SIGTERM)
