+++ cmd/bosun/sched/sched.go
@@ -3,7 +3,6 @@ package sched // import "bosun.org/cmd/bosun/sched"
 import (
 	"encoding/gob"
 	"fmt"
-	"net"
 	"reflect"
 	"strings"
 	"sync"
@@ -25,7 +24,6 @@ import (
 	"github.com/boltdb/bolt"
 	"github.com/bradfitz/slice"
 	"github.com/kylebrandt/boolq"
-	"github.com/tatsushid/go-fastping"
 )
 
 func utcNow() time.Time {
@@ -560,46 +558,6 @@ func Reset() {
 	DefaultSched.Reset()
 }
 
-const pingFreq = time.Second * 15
-
-func (s *Schedule) PingHosts() {
-	for range time.Tick(pingFreq) {
-		hosts, err := s.Search.TagValuesByTagKey("host", s.SystemConf.GetPingDuration())
-		if err != nil {
-			slog.Error(err)
-			continue
-		}
-		for _, host := range hosts {
-			go pingHost(host)
-		}
-	}
-}
-
-func pingHost(host string) {
-	p := fastping.NewPinger()
-	tags := opentsdb.TagSet{"dst_host": host}
-	resolved := 0
-	defer func() {
-		collect.Put("ping.resolved", tags, resolved)
-	}()
-	ra, err := net.ResolveIPAddr("ip4:icmp", host)
-	if err != nil {
-		return
-	}
-	resolved = 1
-	p.AddIPAddr(ra)
-	p.MaxRTT = time.Second * 5
-	timeout := 1
-	p.OnRecv = func(addr *net.IPAddr, t time.Duration) {
-		collect.Put("ping.rtt", tags, float64(t)/float64(time.Millisecond))
-		timeout = 0
-	}
-	if err := p.Run(); err != nil {
-		slog.Errorln(err)
-	}
-	collect.Put("ping.timeout", tags, timeout)
-}
-
 func init() {
 	metadata.AddMetricMeta("bosun.statefile.size", metadata.Gauge, metadata.Bytes,
 		"The total size of the Bosun state file.")
@@ -607,12 +565,7 @@ func init() {
 		"The number of seconds it took Bosun to check each alert rule.")
 	metadata.AddMetricMeta("bosun.check.err", metadata.Gauge, metadata.Error,
 		"The running count of the number of errors Bosun has received while trying to evaluate an alert expression.")
-	metadata.AddMetricMeta("bosun.ping.resolved", metadata.Gauge, metadata.Bool,
-		"1=Ping resolved to an IP Address. 0=Ping failed to resolve to an IP Address.")
-	metadata.AddMetricMeta("bosun.ping.rtt", metadata.Gauge, metadata.MilliSecond,
-		"The number of milliseconds for the echo reply to be received. Also known as Round Trip Time.")
-	metadata.AddMetricMeta("bosun.ping.timeout", metadata.Gauge, metadata.Ok,
-		"0=Ping responded before timeout. 1=Ping did not respond before 5 second timeout.")
+
 	metadata.AddMetricMeta("bosun.actions", metadata.Gauge, metadata.Count,
 		"The running count of actions performed by individual users (Closed alert, Acknowledged alert, etc).")
 }
